[recharts] add to not needed packages, update contributors
