+++ types/recharts/tsconfig.json
@@ -14,11 +14,17 @@
         "typeRoots": [
             "../"
         ],
+        "paths": {
+            "d3-shape": [
+                "d3-shape/v1"
+            ]
+        },
         "types": [],
         "noEmit": true,
         "forceConsistentCasingInFileNames": true
     },
     "files": [
-        "index.d.ts"
+        "index.d.ts",
+        "recharts-tests.tsx"
     ]
 }
