--- types/recharts/v1/tsconfig.json
@@ -1,33 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "lib": [
-            "es6",
-            "dom"
-        ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": true,
-        "baseUrl": "../../",
-        "jsx": "react",
-        "typeRoots": [
-            "../../"
-        ],
-        "paths": {
-            "recharts": [
-                "recharts/v1"
-            ],
-            "d3-shape": [
-                "d3-shape/v1"
-            ]
-        },
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true
-    },
-    "files": [
-        "index.d.ts",
-        "recharts-tests.tsx"
-    ]
-}
